+++ app/src/components/toolbarModals/leaderboard-modal.css
