+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "framer-motion": "^11.0.3"
+  }
+}
